+++ sdk/search/azure-search-documents/src/main/java/com/azure/search/documents/models/SemanticSearchResults.java
@@ -15,6 +15,8 @@ public final class SemanticSearchResults {
     private final List<QueryAnswerResult> queryAnswers;
     private final SemanticErrorReason errorReason;
     private final SemanticSearchResultsType resultsType;
+    private final DebugInfo debugInfo;
+    private final SemanticQueryRewritesResultType semanticQueryRewritesResultType;
 
     static {
         SemanticSearchResultsAccessHelper.setAccessor(SemanticSearchResults::new);
@@ -24,6 +26,8 @@ public final class SemanticSearchResults {
         this.queryAnswers = SearchPagedResponseAccessHelper.getQueryAnswers(pagedResponse);
         this.errorReason = SearchPagedResponseAccessHelper.getSemanticErrorReason(pagedResponse);
         this.resultsType = SearchPagedResponseAccessHelper.getSemanticSearchResultsType(pagedResponse);
+        this.debugInfo = SearchPagedResponseAccessHelper.getDebugInfo(pagedResponse);
+        this.semanticQueryRewritesResultType = SearchPagedResponseAccessHelper.getSemanticQueryRewritesResultType(pagedResponse);
     }
 
     /**
@@ -53,5 +57,23 @@ public final class SemanticSearchResults {
      */
     public SemanticSearchResultsType getResultsType() {
         return this.resultsType;
+    }
+
+    /**
+     * Contains debugging information that can be used to further explore your search results.
+     *
+     * @return The debugging information that can be used to further explore your search results.
+     */
+    public DebugInfo getDebugInfo() {
+        return this.debugInfo;
+    }
+
+    /**
+     * Type of query rewrite that was used for this request.
+     *
+     * @return The type of query rewrite that was used for this request.
+     */
+    public SemanticQueryRewritesResultType getSemanticQueryRewritesResultType() {
+        return this.semanticQueryRewritesResultType;
     }
 }