+++ sdk/search/azure-search-documents/src/samples/java/com/azure/search/documents/SemanticSearchExample.java
@@ -138,6 +138,12 @@ public class SemanticSearchExample {
 
         SemanticSearchResults semanticSearchResults = results.getSemanticResults();
 
+        System.out.println("Semantic Debug Info: " + semanticSearchResults.getDebugInfo());
+
+        System.out.println("Semantic Query Rewrites Result Type: " + semanticSearchResults.getSemanticQueryRewritesResultType());
+
+
+
         System.out.println("Semantic Results Type: " + semanticSearchResults.getResultsType());
 
         if (semanticSearchResults.getErrorReason() != null) {